+++ packages/kiarina-agi-data-builder/src/kiarina/agi/file_segment_normalizer/__init__.py
@@ -3,25 +3,31 @@ from typing import TYPE_CHECKING
 
 if TYPE_CHECKING:
     from ._helpers.normalize_file_segments import normalize_file_segments
-    from ._models.base_file_segment_normalizer import BaseFileSegmentNormalizer
-    from ._operations.create_file_segment_normalizer import (
-        create_file_segment_normalizer,
+    from ._instances.file_segment_normalizer_registry import (
+        file_segment_normalizer_registry,
     )
+    from ._models.base_file_segment_normalizer import BaseFileSegmentNormalizer
     from ._settings import FileSegmentNormalizerSettings, settings_manager
     from ._types.file_segment_normalizer import FileSegmentNormalizer
+    from ._types.file_segment_normalizer_name import FileSegmentNormalizerName
+    from ._types.file_segment_normalizer_specifier import (
+        FileSegmentNormalizerSpecifier,
+    )
 
 __all__ = [
     # ._helpers
     "normalize_file_segments",
+    # ._instances
+    "file_segment_normalizer_registry",
     # ._models
     "BaseFileSegmentNormalizer",
-    # ._operations
-    "create_file_segment_normalizer",
     # ._settings
     "FileSegmentNormalizerSettings",
     "settings_manager",
     # ._types
     "FileSegmentNormalizer",
+    "FileSegmentNormalizerName",
+    "FileSegmentNormalizerSpecifier",
 ]
 
 
@@ -32,15 +38,17 @@ def __getattr__(name: str) -> object:
     module_map = {
         # ._helpers
         "normalize_file_segments": "._helpers.normalize_file_segments",
+        # ._instances
+        "file_segment_normalizer_registry": "._instances.file_segment_normalizer_registry",
         # ._models
         "BaseFileSegmentNormalizer": "._models.base_file_segment_normalizer",
-        # ._operations
-        "create_file_segment_normalizer": "._operations.create_file_segment_normalizer",
         # ._settings
         "FileSegmentNormalizerSettings": "._settings",
         "settings_manager": "._settings",
         # ._types
         "FileSegmentNormalizer": "._types.file_segment_normalizer",
+        "FileSegmentNormalizerName": "._types.file_segment_normalizer_name",
+        "FileSegmentNormalizerSpecifier": "._types.file_segment_normalizer_specifier",
     }
 
     globals()[name] = getattr(import_module(module_map[name], __name__), name)