+++ packages/kiarina-agi-data-builder/src/kiarina/agi/file_segment_normalizer/_instances/file_segment_normalizer_registry.py
@@ -1,0 +1,11 @@
+from kiarina.utils.component_registry import ComponentRegistry
+
+from .._settings import settings_manager
+from .._types.file_segment_normalizer import FileSegmentNormalizer
+
+file_segment_normalizer_registry = ComponentRegistry[FileSegmentNormalizer](
+    expected_type=FileSegmentNormalizer,
+    component_label="FileSegmentNormalizer",
+    get_presets=lambda: settings_manager.settings.presets,
+    get_customs=lambda: settings_manager.settings.customs,
+)