+++ packages/kiarina-agi-data-builder/src/kiarina/agi/file_segment_normalizer/_types/file_segment_normalizer_name.py
@@ -1,0 +1,3 @@
+from typing import TypeAlias
+
+FileSegmentNormalizerName: TypeAlias = str